requirements.txt
langgraph>=0.2.30
langchain>=0.2.16
pydantic>=2.7
python-dateutil>=2.9.0
dateparser>=1.2.0
pandas>=2.2.2
openpyxl>=3.1.5
rich>=13.7.1
click>=8.1.7
uvicorn>=0.30.6
fastapi>=0.115.0
httpx>=0.27.2
modelcontextprotocol>=0.1.4
anthropic>=0.39.0
python-dotenv>=1.0.1
streamlit>=1.38.0
